This is a prototype of an XML scheme for a connection rule. It will be designed to deal with two types of spatial wiring rules.
In the first type of rule the probability of connection is a negative exponential function of distance.

In the second type of rule the probability of allocating a synaptic mechanism to two pieces of adjacent neurite is drawn from a uniform distribution (the probability is constant, it does not vary as a function of space, except for when the distance exceeds a threshol cut_off).